# docs/swagger.yaml
basePath: /api/v1
definitions:
  contracts.BatchDownloadRequest:
    properties:
      auto_classify:
        type: boolean
      directory:
        type: string
      items:
        items:
          $ref: '#/definitions/contracts.DownloadRequest'
        type: array
      video_only:
        type: boolean
    required:
    - items
    type: object
...
  contracts.DownloadRequest:
    properties:
      auto_classify:
        type: boolean
      directory:
        type: string
      file_size:
        description: 文件大小，用于磁盘空间检查
        type: integer
      filename:
        type: string
      options:
        additionalProperties: true
        type: object
      url:
        type: string
      video_only:
        type: boolean
    required:
    - url
    type: object
...
  contracts.FileListRequest:
    properties:
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 1000
        minimum: 1
        type: integer
      path:
        type: string
      recursive:
        type: boolean
      sort_by:
        enum:
        - name
        - size
        - modified
        type: string
      sort_order:
        enum:
        - asc
        - desc
        type: string
      video_only:
        type: boolean
    required:
    - path
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于Gin框架的Alist和Aria2下载管理服务
  license:
    name: MIT
    url: https://github.com/easayliu/alist-aria2-download/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Alist Aria2 Download API
  version: "1.0"
paths:
  /alist/file:
    get:
      consumes:
      - application/json
      description: 获取Alist中指定路径文件的详细信息，包含下载链接
      parameters:
      - description: 文件完整路径
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件详细信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件信息
      tags:
      - Alist管理
  /alist/files:
    get:
      consumes:
      - application/json
      description: 获取Alist中指定路径的文件列表
      parameters:
      - description: 文件路径（留空使用配置的默认路径）
        in: query
        name: path
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件列表
      tags:
      - Alist管理
  /alist/login:
    post:
      description: 使用配置文件中的用户名密码登录Alist服务
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 登录失败
          schema:
            additionalProperties: true
            type: object
      summary: Alist登录
      tags:
      - Alist管理
  /downloads:
    get:
      description: 获取所有Aria2下载任务列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取下载列表
      tags:
      - 下载管理
    post:
      consumes:
      - application/json
      description: 创建新的Aria2下载任务
      parameters:
      - description: 下载请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.DownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 下载任务创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建下载任务
      tags:
      - 下载管理
  /downloads/{id}:
    delete:
      description: 根据GID删除下载任务
      parameters:
      - description: 下载任务GID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 删除下载任务
      tags:
      - 下载管理
    get:
      description: 根据GID获取单个下载任务详情
      parameters:
      - description: 下载任务GID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取下载详情
      tags:
      - 下载管理
  /downloads/{id}/pause:
    post:
      description: 暂停指定的下载任务
      parameters:
      - description: 下载任务GID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 暂停下载
      tags:
      - 下载管理
  /downloads/{id}/resume:
    post:
      description: 恢复指定的下载任务
      parameters:
      - description: 下载任务GID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 恢复下载
      tags:
      - 下载管理
  /downloads/batch:
    post:
      consumes:
      - application/json
      description: 批量创建多个Aria2下载任务
      parameters:
      - description: 批量下载请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.BatchDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量下载任务创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 批量创建下载任务
      tags:
      - 下载管理
  /downloads/pause-all:
    post:
      description: 暂停所有正在进行的下载任务
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 暂停所有下载
      tags:
      - 下载管理
  /downloads/resume-all:
    post:
      description: 恢复所有已暂停的下载任务
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 恢复所有下载
      tags:
      - 下载管理
  /downloads/statistics:
    get:
      description: 获取下载任务的统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取下载统计
      tags:
      - 下载管理
  /downloads/system-status:
    get:
      description: 获取Aria2系统的状态信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取系统状态
      tags:
      - 下载管理
...
  /files/list:
    post:
      consumes:
      - application/json
      description: 获取指定路径下的文件列表，支持分页和视频文件过滤
      parameters:
      - description: 列出文件请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.FileListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 列出指定路径的文件
      tags:
      - 文件管理
...
schemes:
- http
- https
swagger: "2.0"